ark-core/autoscaler/requirements.txt
nats-py==2.9.0
docker==7.1.0
